+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## unreleased
 - veproduct.Type: add IsBMV, IsSolar, and IsInverter methods.
+- veproduct.Product: add MaxPanelVoltage and MaxPanelCurrent methods.